Disable the puppetserver module calls
The puppetserver module is not properly compatible with puppet4
currently. While it doesn't cause it to outright fail to work, because
of it not being able to manage puppetserver gems currently it tosses
errors that get picked up by the OpsGenie reporter.

This disables the calls into the module for now.

Change-Id: If2c3008939b776c36f794a4b52f04f6bd06cad4e
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>