Convert hiera lookup for bacula::jobs to merge
The default hiera lookup does not do hash merges for hiera entries.
Given that domains are likely to want to have a base set of jobs that
all systems get. Configuring hiera to do hash merges allows nodes to add
extra jobs that need to happen on top of the base jobs.

NOTE: if the node defines a job that exists at a lower level, the nodes
version will take precedence!

Change-Id: I3539feb91bd3babef502ec4e407c2b9f5917d7f2
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>