Add in initial nagios profiles
This adds in the initial nagios profiles. It has the calls to ::nagios
(for server) and ::nagios::client (for clients) currently commented out
since we don't have the nagios module currently defined in the master
repo.

What this really does is prep us for when it _is_ defined and also let
us start configuring up the NRPE commands.

DEPENDS on Gerrit 367 -
https://gerrit.linuxfoundation.org/infra/#/c/367/

Change-Id: Ie45602fe58f16e4b6ab0fade592edb5fa47cb56c
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>