Create NRPE commands with more dynamic config
We've got a couple of NRPE commands that need to be a little bit more
dynamic in their configuration.

The calculated values in use are the ones that were (are?) in the legacy
CAF puppet system. They aren't exactly what LFCore has been using. If we
feel that we need these to be a little more tunable to the different
environments we'll need to push some of the calculation values into
hiera and load them up to then do the calculations.

Change-Id: Ib1f7187eb998ba11f7f741fab287ffde6ef6db0e
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>