Migrate base role to a base profile
The recent Puppet training made me realize that we should be doing a
slightly different strategy for our base configuration. It should not be
a base "role" so much as a base profile.

This is the first step in migrating from doing the base configuration in
a profile and subsequently doing away with the role inheritance that
we're currently using and moving to a more appropriate profile include
in all roles.

Change-Id: If9cb35244fdd138099cfe25b7baed272a9386667
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>