Force postfix transport maps to update
On a clean install on EL systems postfix pushes out a commented
transport map. Puppet doesn't handle this very well with the transport
map updates that it should do and so mail doesn't work. This causes us
to throw away the transport map file if it matches the default RPM
version which should only be on initial setup.

Change-Id: I91d568b1f324d6abafc3b8e9e2671bd46372f711
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>