Only execute report tidy operation once a day
Having the puppet master spam out notices via tagmail (if it's
configured) every run because of report tidy operations is a bit
annoying and well, spammy. This reconfigures the tidy operation to only
happen once per day but still honors the report_ttl configuration though
because of this it does mean that the granularity for tidy operations is
now as the smallest, 1 day. Which in all honesty should be fine!

Change-Id: I82426e40d74177b5ccc27b1f1cabf53b71b81bdf
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>